# internal/server/openapi.yaml
openapi: 3.1.0
info:
  title: PaiHuo API
  version: 1.0.0
  description: PaiHuo control-plane contract. Browser sessions use an HttpOnly cookie obtained from POST /login.
servers:
  - url: /api/v1
...
paths:
...
  /sessions:
    get: {tags: [Sessions], responses: {'200': {description: Sessions}}}
    post: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: {'201': {description: Session created}}}
  /sessions/{id}:
    get: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Session}}}
    delete: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {description: Session deleted}}}
  /sessions/{id}/start:
    post: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {description: Started}}}
  /sessions/{id}/resume:
    post: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {description: Resumed}}}
  /sessions/{id}/suspend:
    post: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {description: Suspended}}}
  /sessions/{id}/prompt:
    post: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {description: Prompt accepted}}}
...
  /sessions/{id}/transcript:
    get: {tags: [Sessions], parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Transcript page}}}
  /workflow-proposals:
    get: {tags: [Workflows], responses: {'200': {description: Proposals}}}
    post: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/IdempotencyKey'}], responses: {'201': {description: Proposal created}}}
  /workflow-proposals/{id}:
    get: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Proposal}}}
  /workflow-proposals/{id}/validate:
    post: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Revision'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'200': {description: Policy result}}}
  /workflow-proposals/{id}/adopt:
    post: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Revision'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'201': {description: Frozen plan}}}
...
  /workflows/{id}:
    get: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Frozen plan}}}
  /workflows/{id}/runs:
    post: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/ID'}, {$ref: '#/components/parameters/Revision'}, {$ref: '#/components/parameters/IdempotencyKey'}], responses: {'201': {description: Workflow run created atomically}}}
  /workflow-runs/{id}:
    get: {tags: [Workflows], parameters: [{$ref: '#/components/parameters/ID'}], responses: {'200': {description: Workflow run}}}
...
components:
  parameters:
    ID: {name: id, in: path, required: true, schema: {type: integer, format: int64, minimum: 1}}
    Revision: {name: If-Match, in: header, required: true, description: Quoted revision returned in the resource body or ETag, schema: {type: string, example: '"3"'}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, description: Replays the first completed response for the same method and path, schema: {type: string, maxLength: 200}}
...